+++ src/kili/domain/asset/__init__.py
@@ -1,5 +1,5 @@
 """Asset domain."""
 
-from .asset import AssetExternalId, AssetFilters, AssetId
+from .asset import AssetExternalId, AssetFilters, AssetId, AssetStatus
 
-__all__ = ["AssetFilters", "AssetId", "AssetExternalId"]
+__all__ = ["AssetFilters", "AssetId", "AssetExternalId", "AssetStatus"]
